+++ pynumaflow/function/sync_server.py
@@ -1,28 +1,21 @@
-from concurrent.futures import ThreadPoolExecutor
-
-import asyncio
-import grpc
 import logging
 import multiprocessing
 import os
-from datetime import datetime, timezone
+from concurrent.futures import ThreadPoolExecutor
+from datetime import datetime
+from typing import Callable, AsyncIterable, List
+
+import grpc
 from google.protobuf import empty_pb2 as _empty_pb2
 from google.protobuf import timestamp_pb2 as _timestamp_pb2
-from typing import Callable, AsyncIterable, List
 
 from pynumaflow import setup_logging
 from pynumaflow._constants import (
     FUNCTION_SOCK_PATH,
-    WIN_START_TIME,
-    WIN_END_TIME,
     MAX_MESSAGE_SIZE,
-    STREAM_EOF,
-    DELIMITER,
 )
-from pynumaflow.function import Messages, MessageTs, Datum, IntervalWindow, Metadata
-from pynumaflow.function._dtypes import ReduceResult, DatumMetadata
-from pynumaflow.function._multiproc_server import MultiProcServer
-from pynumaflow.function.asynciter import NonBlockingIterator
+from pynumaflow.function import Messages, MessageTs, Datum, Metadata
+from pynumaflow.function._dtypes import DatumMetadata
 from pynumaflow.function.proto import udfunction_pb2
 from pynumaflow.function.proto import udfunction_pb2_grpc
 from pynumaflow.types import NumaflowServicerContext
@@ -38,24 +31,7 @@ _PROCESS_COUNT = multiprocessing.cpu_count()
 MAX_THREADS = int(os.getenv("MAX_THREADS", 0)) or (_PROCESS_COUNT * 4)
 
 
-async def datum_generator(
-    request_iterator: AsyncIterable[udfunction_pb2.DatumRequest],
-) -> AsyncIterable[Datum]:
-    async for d in request_iterator:
-        datum = Datum(
-            keys=list(d.keys),
-            value=d.value,
-            event_time=d.event_time.event_time.ToDatetime(),
-            watermark=d.watermark.watermark.ToDatetime(),
-            metadata=DatumMetadata(
-                msg_id=d.metadata.id,
-                num_delivered=d.metadata.num_delivered,
-            ),
-        )
-        yield datum
-
-
-class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionServicer):
+class SyncServer(udfunction_pb2_grpc.UserDefinedFunctionServicer):
     """
     Provides an interface to write a User Defined Function (UDFunction)
     which will be exposed over gRPC.
@@ -72,8 +48,7 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
     Example invocation:
     >>> from typing import Iterator
     >>> from pynumaflow.function import Messages, Message, MessageTs, MessageT, \
-    ...     Datum, Metadata, UserDefinedFunctionServicer
-    ... import aiorun
+    ...     Datum, Metadata, SyncServer
     ...
     >>> def map_handler(key: [str], datum: Datum) -> Messages:
     ...   val = datum.value
@@ -100,22 +75,22 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
     ...   )
     ...   return Messages(Message.to_vtx(key, str.encode(msg)))
     ...
-    >>> grpc_server = UserDefinedFunctionServicer(
+    >>> grpc_server = SyncServer(
     ...   reduce_handler=reduce_handler,
     ...   mapt_handler=mapt_handler,
     ...   map_handler=map_handler,
     ... )
-    >>> aiorun.run(grpc_server.start_async())
+    >>> aiorun.run(grpc_server.start())
     """
 
     def __init__(
-        self,
-        map_handler: UDFMapCallable = None,
-        mapt_handler: UDFMapTCallable = None,
-        reduce_handler: UDFReduceCallable = None,
-        sock_path=FUNCTION_SOCK_PATH,
-        max_message_size=MAX_MESSAGE_SIZE,
-        max_threads=MAX_THREADS,
+            self,
+            map_handler: UDFMapCallable = None,
+            mapt_handler: UDFMapTCallable = None,
+            reduce_handler: UDFReduceCallable = None,
+            sock_path=FUNCTION_SOCK_PATH,
+            max_message_size=MAX_MESSAGE_SIZE,
+            max_threads=MAX_THREADS,
     ):
         if not (map_handler or mapt_handler or reduce_handler):
             raise ValueError("Require a valid map/mapt handler and/or a valid reduce handler.")
@@ -138,7 +113,7 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
         ]
 
     def MapFn(
-        self, request: udfunction_pb2.DatumRequest, context: NumaflowServicerContext
+            self, request: udfunction_pb2.DatumRequest, context: NumaflowServicerContext
     ) -> udfunction_pb2.DatumResponseList:
         """
         Applies a function to each datum element.
@@ -174,7 +149,7 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
         return udfunction_pb2.DatumResponseList(elements=datums)
 
     def MapTFn(
-        self, request: udfunction_pb2.DatumRequest, context: NumaflowServicerContext
+            self, request: udfunction_pb2.DatumRequest, context: NumaflowServicerContext
     ) -> udfunction_pb2.DatumResponseList:
         """
         Applies a function to each datum element.
@@ -218,141 +193,26 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
             )
         return udfunction_pb2.DatumResponseList(elements=datums)
 
-    async def ReduceFn(
-        self,
-        request_iterator: AsyncIterable[udfunction_pb2.DatumRequest],
-        context: NumaflowServicerContext,
-    ) -> udfunction_pb2.DatumResponseList:
-        """
-        Applies a reduce function to a datum stream.
-        The pascal case function name comes from the proto udfunction_pb2_grpc.py file.
-        """
-
-        start, end = None, None
-        for metadata_key, metadata_value in context.invocation_metadata():
-            if metadata_key == WIN_START_TIME:
-                start = metadata_value
-            elif metadata_key == WIN_END_TIME:
-                end = metadata_value
-        if not (start or end):
-            raise ValueError(
-                f"Expected to have all key/window_start_time/window_end_time; "
-                f"got start: {start}, end: {end}."
-            )
-
-        start_dt = datetime.fromtimestamp(int(start) / 1e3, timezone.utc)
-        end_dt = datetime.fromtimestamp(int(end) / 1e3, timezone.utc)
-        interval_window = IntervalWindow(start=start_dt, end=end_dt)
-
-        datum_iterator = datum_generator(request_iterator=request_iterator)
-
-        response_task = asyncio.create_task(
-            self.__async_reduce_handler(interval_window, datum_iterator)
-        )
-
-        # Save a reference to the result of this function, to avoid a
-        # task disappearing mid-execution.
-        self.background_tasks.add(response_task)
-        response_task.add_done_callback(lambda t: self.background_tasks.remove(t))
-
-        await response_task
-        results_futures = response_task.result()
-
-        for fut in results_futures:
-            await fut
-            yield udfunction_pb2.DatumResponseList(elements=fut.result())
-
-    async def __async_reduce_handler(self, interval_window, datum_iterator: AsyncIterable[Datum]):
-        callable_dict = {}
-        # iterate through all the values
-        async for d in datum_iterator:
-            keys = d.keys()
-            unified_key = DELIMITER.join(keys)
-            result = callable_dict.get(unified_key, None)
-
-            if not result:
-                niter = NonBlockingIterator()
-                riter = niter.read_iterator()
-                # schedule an async task for consumer
-                # returns a future that will give the results later.
-                task = asyncio.create_task(
-                    self.__invoke_reduce(keys, riter, Metadata(interval_window=interval_window))
-                )
-                # Save a reference to the result of this function, to avoid a
-                # task disappearing mid-execution.
-                self.background_tasks.add(task)
-                task.add_done_callback(lambda t: self.background_tasks.remove(t))
-                result = ReduceResult(task, niter, keys)
-
-                callable_dict[unified_key] = result
-
-            await result.iterator.put(d)
-
-        for unified_key in callable_dict:
-            await callable_dict[unified_key].iterator.put(STREAM_EOF)
-
-        tasks = []
-        for unified_key in callable_dict:
-            fut = callable_dict[unified_key].future
-            tasks.append(fut)
-
-        return tasks
-
-    async def __invoke_reduce(
-        self, keys: List[str], request_iterator: AsyncIterable[Datum], md: Metadata
-    ):
-        try:
-            msgs = await self.__reduce_handler(keys, request_iterator, md)
-        except Exception as err:
-            _LOGGER.critical("UDFError, re-raising the error: %r", err, exc_info=True)
-            raise err
-
-        datum_responses = []
-        for msg in msgs.items():
-            datum_responses.append(
-                udfunction_pb2.DatumResponse(keys=msg.keys, value=msg.value, tags=msg.tags)
-            )
-
-        return datum_responses
+    # def ReduceFn(
+    #         self,
+    #         request_iterator: AsyncIterable[udfunction_pb2.Datum],
+    #         context: NumaflowServicerContext,
+    # ) -> udfunction_pb2.DatumList:
+    #     """
+    #     Applies a reduce function to a datum stream.
+    #     The pascal case function name comes from the proto udfunction_pb2_grpc.py file.
+    #     """
+    #     _LOGGER.error("Reduce not supported on NEW SYNC --")
+    #     raise NotImplementedError("Reduce Not supported on sync")
 
     def IsReady(
-        self, request: _empty_pb2.Empty, context: NumaflowServicerContext
+            self, request: _empty_pb2.Empty, context: NumaflowServicerContext
     ) -> udfunction_pb2.ReadyResponse:
         """
         IsReady is the heartbeat endpoint for gRPC.
         The pascal case function name comes from the proto udfunction_pb2_grpc.py file.
         """
         return udfunction_pb2.ReadyResponse(ready=True)
-
-    async def __serve_async(self, server) -> None:
-        udfunction_pb2_grpc.add_UserDefinedFunctionServicer_to_server(
-            UserDefinedFunctionServicer(
-                map_handler=self.__map_handler,
-                mapt_handler=self.__mapt_handler,
-                reduce_handler=self.__reduce_handler,
-            ),
-            server,
-        )
-        server.add_insecure_port(self.sock_path)
-        _LOGGER.info("GRPC Async Server listening on: %s", self.sock_path)
-        await server.start()
-
-        async def server_graceful_shutdown():
-            """
-            Shuts down the server with 5 seconds of grace period. During the
-            grace period, the server won't accept new connections and allow
-            existing RPCs to continue within the grace period.
-            """
-            _LOGGER.info("Starting graceful shutdown...")
-            await server.stop(5)
-
-        self.cleanup_coroutines.append(server_graceful_shutdown())
-        await server.wait_for_termination()
-
-    async def start_async(self) -> None:
-        """Starts the Async gRPC server on the given UNIX socket."""
-        server = grpc.aio.server(options=self._server_options)
-        await self.__serve_async(server)
 
     def start(self) -> None:
         """
@@ -368,10 +228,3 @@ class UserDefinedFunctionServicer(udfunction_pb2_grpc.UserDefinedFunctionService
             "GRPC Server listening on: %s with max threads: %s", self.sock_path, self._max_threads
         )
         server.wait_for_termination()
-
-    def start_multiproc(self) -> None:
-        """
-        Starts multiple gRPC servers, all bound on a given TCP socket.
-        """
-        server = MultiProcServer(udf_service=self, server_options=self._server_options)
-        server.start()
